+++ packages/oauth/oauth-provider/src/account/account-manager.ts
@@ -55,20 +55,30 @@ export class AccountManager {
       throw new InvalidRequestError('hCaptcha token is required')
     }
 
-    const { allowed, result } = await this.hcaptchaClient
-      .verify(
-        'signup',
-        input.hcaptchaToken,
-        deviceMetadata.ipAddress,
-        input.handle,
-        deviceMetadata.userAgent,
-      )
+    const tokens = this.hcaptchaClient.buildClientTokens(
+      deviceMetadata.ipAddress,
+      input.handle,
+      deviceMetadata.userAgent,
+    )
+
+    const result = await this.hcaptchaClient
+      .verify('signup', input.hcaptchaToken, deviceMetadata.ipAddress, tokens)
       .catch((err) => {
         throw InvalidRequestError.from(err, 'hCaptcha verification failed')
       })
 
-    if (!allowed) {
-      throw new InvalidRequestError('hCaptcha verification failed')
+    await callAsync(this.hooks.onHcaptchaResult, {
+      input,
+      deviceId,
+      deviceMetadata,
+      tokens,
+      result,
+    })
+
+    try {
+      this.hcaptchaClient.checkVerifyResult(result, tokens)
+    } catch (err) {
+      throw InvalidRequestError.from(err, 'hCaptcha verification failed')
     }
 
     return result